chore: bump dev deps and refactor usages
