+++ package.json
@@ -17,8 +17,8 @@
     "generate:css": "tailwindcss -o ./app/styles/tailwind.css",
     "lint": "eslint --cache --cache-location ./node_modules/.cache/eslint .",
     "test": "vitest",
-    "test:e2e:dev": "start-server-and-test dev http://localhost:3000 \"cypress open\"",
-    "test:e2e:run": "cross-env PORT=8811 start-server-and-test dev http://localhost:8811 \"cypress run\"",
+    "test:e2e:dev": "start-server-and-test dev http://localhost:3000 \"npx cypress open\"",
+    "test:e2e:run": "cross-env PORT=8811 start-server-and-test dev http://localhost:8811 \"npx cypress run\"",
     "typecheck": "tsc -b && tsc -b cypress"
   },
   "dependencies": {
@@ -40,33 +40,33 @@
     "@faker-js/faker": "^6.3.1",
     "@remix-run/dev": "*",
     "@remix-run/eslint-config": "*",
-    "@testing-library/cypress": "^8.0.2",
-    "@testing-library/jest-dom": "^5.16.4",
-    "@testing-library/react": "^13.2.0",
-    "@testing-library/user-event": "^14.2.0",
-    "@types/architect__functions": "^3.13.6",
+    "@testing-library/cypress": "^8.0.3",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^14.4.3",
+    "@types/architect__functions": "^3.13.7",
     "@types/bcryptjs": "2.4.2",
-    "@types/eslint": "^8.4.2",
-    "@types/ramda": "^0.28.13",
-    "@types/react": "^18.0.9",
-    "@types/react-dom": "^18.0.4",
-    "@vitejs/plugin-react": "^1.3.2",
+    "@types/eslint": "^8.4.6",
+    "@types/ramda": "^0.28.15",
+    "@types/react": "^18.0.20",
+    "@types/react-dom": "^18.0.6",
+    "@vitejs/plugin-react": "^2.1.0",
     "@vitest/coverage-c8": "^0.23.2",
-    "c8": "^7.11.3",
+    "c8": "^7.12.0",
     "cross-env": "^7.0.3",
-    "cypress": "^9.7.0",
-    "esbuild": "^0.14.39",
-    "eslint": "^8.16.0",
+    "cypress": "^10.8.0",
+    "esbuild": "^0.15.7",
+    "eslint": "^8.23.1",
     "eslint-config-prettier": "^8.5.0",
-    "happy-dom": "^4.0.1",
-    "msw": "^0.41.0",
+    "happy-dom": "^6.0.4",
+    "msw": "^0.47.3",
     "npm-run-all": "^4.1.5",
-    "prettier": "^2.6.2",
-    "prettier-plugin-tailwindcss": "^0.1.11",
+    "prettier": "^2.7.1",
+    "prettier-plugin-tailwindcss": "^0.1.13",
     "start-server-and-test": "^1.14.0",
-    "tailwindcss": "^3.0.24",
-    "ts-node": "^10.8.0",
-    "typescript": "^4.6.4",
+    "tailwindcss": "^3.1.8",
+    "ts-node": "^10.9.1",
+    "typescript": "^4.8.3",
     "vite-tsconfig-paths": "^3.5.0",
     "vitest": "^0.23.2"
   },
